Diagonal Difference in JAVA
